requirements.txt
Click==7.0
cycler==0.10.0
Cython==0.29.13
decorator==4.4.0
Flask==1.1.1
gunicorn==19.9.0
imageio==2.5.0
itsdangerous==1.1.0
Jinja2==2.10.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
networkx==2.3
numpy==1.17.2
opencv-contrib-python==4.1.1.26
opencv-python-headless==4.1.1.26
Pillow==6.1.0
pyocr==0.7.2
pyparsing==2.4.2
pytesseract==0.3.0
python-dateutil==2.8.0
PyWavelets==1.0.3
scikit-image==0.15.0
scipy==1.3.1
six==1.12.0
Werkzeug==0.16.0
